test(server): enforce workspace Git ownership
Parallel Workspace tests that spawn clones and worktrees were still concentrated in the root server binary without an independent concurrency bound. That let unrelated root tests amplify Git process pressure and obscured which cases genuinely exercise composition.

Keep handler-owned lifecycle and clone-safety coverage in the weighted Workspace lane, and bound the smaller provider, config, Kata, runtime, and agent-context composition set with a root-only semaphore.
